erc721, constants, events, and primitive data codes updated and checked
