+++ example_code/applications/erc721/Cargo.toml
@@ -1,17 +1,32 @@
 [package]
-name = "stylus-erc721"
-version = "0.1.0"
+name = "stylus_erc721_example"
+version = "0.1.7"
 edition = "2021"
 license = "MIT OR Apache-2.0"
+homepage = "https://github.com/OffchainLabs/stylus-hello-world"
+repository = "https://github.com/OffchainLabs/stylus-hello-world"
+keywords = ["arbitrum", "ethereum", "stylus", "alloy"]
+description = "Stylus hello world example"
 
 [dependencies]
-stylus-sdk = "0.5.0"
-mini-alloc = "0.4.2"
 alloy-primitives = "0.7.6"
 alloy-sol-types = "0.7.6"
+mini-alloc = "0.4.2"
+stylus-sdk = "0.5.2"
+hex = "0.4.3"
+
+[dev-dependencies]
+tokio = { version = "1.12.0", features = ["full"] }
+ethers = "2.0"
+eyre = "0.6.8"
 
 [features]
 export-abi = ["stylus-sdk/export-abi"]
+debug = ["stylus-sdk/debug"]
+
+[[bin]]
+name = "stylus-hello-world"
+path = "src/main.rs"
 
 [lib]
 crate-type = ["lib", "cdylib"]
