+++ example_code/applications/erc721/src/erc721.rs
@@ -168,11 +168,11 @@ impl<T: Erc721Params> Erc721<T> {
         if to.has_code() {
             let receiver = IERC721TokenReceiver::new(to);
             let received = receiver
-                .on_erc_721_received(&mut *storage, msg::sender(), from, token_id, data)
+                .on_erc_721_received(&mut *storage, msg::sender(), from, token_id, data.into())
                 .map_err(|_e| Erc721Error::ReceiverRefused(ReceiverRefused {
                     receiver: receiver.address,
                     token_id,
-                    returned: 0_u32.to_be_bytes(),
+                    returned: alloy_primitives::FixedBytes(0_u32.to_be_bytes()),
                 }))?
                 .0;
 
@@ -180,7 +180,7 @@ impl<T: Erc721Params> Erc721<T> {
                 return Err(Erc721Error::ReceiverRefused(ReceiverRefused {
                     receiver: receiver.address,
                     token_id,
-                    returned: received,
+                    returned: alloy_primitives::FixedBytes(received),
                 }));
             }
         }
