+++ example_code/applications/erc721/src/main.rs
@@ -1,4 +1,4 @@
 #[cfg(feature = "export-abi")]
 fn main() {
-    stylus_erc721::print_abi("MIT-OR-APACHE-2.0", "pragma solidity ^0.8.23;");
+    stylus_erc721_example::print_abi("MIT-OR-APACHE-2.0", "pragma solidity ^0.8.23;");
 }
